requirements.txt
Pillow
opencv-python
albumentations
